[Feat] #23 AuthLogoutTest
